Update ht_product_schema.php
Change ratingValue to a float instead of interger.  In testing a 4.9 average gets downgraded to a 4.0 using interger.  With float it shows 4.9.

+++ includes/modules/header_tags/ht_product_schema.php
@@ -98,7 +98,7 @@
           $average = tep_db_fetch_array($average_query);
           if ($average['count'] > 0) {
             $schema_product['aggregateRating'] = array("@type"       => "AggregateRating",
-                                                       "ratingValue" => (int)$average['average'],
+                                                       "ratingValue" => (float)$average['average'],
                                                        "reviewCount" => (int)$average['count']);
           }
